src/requirements.txt
alembic==1.13.3
bcrypt==4.2.0
blinker==1.8.2
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cssselect2==0.7.0
Flask==2.2.5
Flask-Admin==1.6.1
Flask-Cors==5.0.0
Flask-JWT-Extended==4.3.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-SQLAlchemy==2.5.1
flask-swagger==0.2.14
Flask-WTF==1.2.1
fonttools==4.54.1
greenlet==3.1.1
gunicorn==23.0.0
html5lib==1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
packaging==24.2
pillow==10.4.0
psycopg2-binary==2.9.6
pycparser==2.22
pydyf==0.11.0
PyJWT==2.9.0
pyphen==0.16.0
python-dotenv==1.0.1
PyYAML==6.0.2
reportlab==4.2.5
requests==2.32.3
six==1.16.0
SQLAlchemy==1.4.46
stripe==11.1.0
tinycss2==1.3.0
typing_extensions==4.12.2
urllib3==2.2.3
weasyprint==62.3
webencodings==0.5.1
Werkzeug==3.0.6
WTForms==3.1.2
zopfli==0.2.3
flask-talisman==1.1.0
python-slugify==8.0.1
